third_party/WebKit/LayoutTests/intersection-observer/observer-without-js-reference-expected.txt

IntersectionObserver continues to produce notifications when it has no javascript references.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS entries.length is 1
PASS successfullyParsed is true

TEST COMPLETE
